Frog 1.0
The problem I'm having is that he will often do like a jump back and forth instead of going his whole length. It's actually kind of nice because it's is set up that it at least looks mostly right, but I'd like to be able to control exactly what he does instead of being happy he can go randomly.